requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
motor==2.1.0
multidict==4.7.6
pymongo==3.10.1
python-baseconv==1.2.2
yarl==1.4.2
